+++ src/main/java/net/lustenauer/utils/strings/STR.java
@@ -1,7 +1,7 @@
 /*
- * The MIT License (MIT)
+ * MIT License
  *
- * Copyright (c) 2017 Patric Hollenstein
+ * Copyright (c) 2017 - 2018 Patric Hollenstein
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,17 +20,22 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
+ *
  */
 
 package net.lustenauer.utils.strings;
 
 /**
- * Some String methodes
+ * Some String methods
  *
  * @author Patric Hollenstein
  * @since 1.0.0
  */
-public class STR {
+public final class STR {
+
+    /*
+     * PUBLIC METHODS
+     */
 
     /**
      * Similar to substring from the {@link String} class but with more error handling. It is also possible to set
@@ -40,7 +45,6 @@ public class STR {
      * @param start start value from the substring
      * @param end   end value from the substring
      * @return the wanted part of the string or a blank string when the given values out of range
-     * @version 1.0.0
      * @since 1.0.0
      */
     public static String substr(String value, int start, int end) {
@@ -51,5 +55,75 @@ public class STR {
         }
 
         return value.substring(start, end);
+    }
+
+    /**
+     * Performs single argument substitution for the 'pattern' passed as parameter.
+     * <p>
+     * For example,
+     * <p>
+     * <pre>
+     * STR.form(&quot;Hi {}.&quot;, &quot;there&quot;);
+     * </pre>
+     * <p>
+     * will return the string "Hi there.".
+     * <p>
+     *
+     * @param pattern The pattern which will be parsed and formatted
+     * @param arg     The argument to be substituted in place of the formatting anchor
+     * @return The formatted string
+     * @since 1.2.0
+     */
+    public static String form(String pattern, Object arg) {
+        return Formatter.arrayFormat(pattern, new Object[]{arg});
+    }
+
+    /**
+     * Performs a two argument substitution for the 'pattern' passed as parameter.
+     * <p>
+     * For example,
+     * <p>
+     * <pre>
+     * STR.form(&quot;Hi {}. My name is {}.&quot;, &quot;Alice&quot;, &quot;Bob&quot;);
+     * </pre>
+     * <p>
+     * will return the string "Hi Alice. My name is Bob.".
+     *
+     * @param pattern The pattern which will be parsed and formatted
+     * @param arg1    The first argument to be substituted in place of the formatting anchor
+     * @param arg2    The second argument to be substituted in place of the formatting anchor
+     * @return The formatted string
+     * @since 1.2.0
+     */
+    public static String form(String pattern, Object arg1, Object arg2) {
+        return Formatter.arrayFormat(pattern, new Object[]{arg1, arg2});
+    }
+
+    /**
+     * Performs a multiple argument substitution for the 'pattern' passed as parameter.
+     * <p>
+     * For example,
+     * <p>
+     * <pre>
+     * STR.form(&quot;Hi {}. My name is {} and i am {} years old.&quot;, &quot;Alice&quot;, &quot;Bob&quot, 27;);
+     * </pre>
+     * <p>
+     * will return the string "Hi Alice. My name is Bob and i am 27 years old.".
+     *
+     * @param pattern The pattern which will be parsed and formatted
+     * @param args    The arguments to be substituted in place of the formatting anchor
+     * @return The formatted string
+     * @since 1.2.0
+     */
+    public static String form(String pattern, Object... args) {
+        return Formatter.arrayFormat(pattern, args);
+    }
+
+
+    /*
+     * CONSTRUCTORS
+     */
+
+    private STR() {
     }
 }